+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -185,6 +185,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     sonicsvm: true,
     soon: true,
     sophon: true,
+    starknet: true,
     story: true,
     stride: false,
     subtensor: true,
@@ -334,6 +335,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     sonicsvm: true,
     soon: true,
     sophon: true,
+    starknet: true,
     story: true,
     stride: true,
     subtensor: true,
@@ -483,6 +485,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     sonicsvm: true,
     soon: true,
     sophon: true,
+    starknet: true,
     story: true,
     stride: true,
     subtensor: true,
@@ -596,6 +599,7 @@ const gasPaymentEnforcement: GasPaymentEnforcement[] = [
         originDomain: getDomainId('infinityvmmainnet'),
         destinationDomain: getDomainId('solanamainnet'),
       },
+      { originDomain: getDomainId('starknet') },
     ],
   },
   {
@@ -799,7 +803,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: 'b916cd9-20250527-145115',
+      tag: 'b431cda-20250529-133433',
     },
     blacklist,
     gasPaymentEnforcement: gasPaymentEnforcement,
@@ -813,7 +817,7 @@ const hyperlane: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: 'b916cd9-20250527-145115',
+      tag: '2903fe1-20250602-134907',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -824,7 +828,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '197fd87-20250521-144619',
+      tag: '2903fe1-20250602-134907',
     },
     resources: scraperResources,
   },
@@ -839,7 +843,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '4c678c5-20250521-125729',
+      tag: '2903fe1-20250602-134907',
     },
     blacklist,
     // We're temporarily (ab)using the RC relayer as a way to increase
@@ -877,7 +881,7 @@ const neutron: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '0fffbfb-20250514-130038',
+      tag: 'b431cda-20250529-133433',
     },
     blacklist,
     gasPaymentEnforcement,